+++ korus-ui/components/DropZone/hooks.ts
@@ -5,9 +5,9 @@ import {
 } from './types';
 import { useElement, useProps } from '../../utils';
 import { Div } from '../Div';
-import { Ul } from '../Ul';
 import { Span } from '../Span';
-import { Button } from '../Button';
+import { Ul } from '../Ul';
+import { A } from '../A';
 import { LedaContext } from '../LedaProvider';
 import { COMPONENTS_NAMESPACES } from '../../constants';
 
@@ -58,7 +58,7 @@ export const useCustomElements = (props: DropZoneProps, state: DropZoneState): C
 
   const UploadButton = useElement<DropZoneProps, DropZoneState, UploadButtonProps>(
     'UploadButton',
-    Button,
+    A,
     uploadButtonRender || context.renders[COMPONENTS_NAMESPACES.dropZone].uploadButtonRender,
     props,
     state,