+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - MultiSelect: Added header and footer in checkboxes list
 - MultiSelect: Refactoring
 - FileList: Added new component
+- DropZone: Add new styles
 
 ## [0.45.1] - 2021-04-29